Extract output buffering and getopt builtins from vm.c
Move two self-contained subsystems into dedicated files:

- vm_builtin_ob.c (431 lines): ob_start, ob_flush, ob_get_contents,
  and 8 other output buffering functions with their helpers
- vm_builtin_getopt.c (320 lines): getopt() and CLI option parsing

VmObEntry struct moved to ph7int.h (needed by both vm.c init and
the OB module). VmObConsumer promoted to PH7_PRIVATE since it's used
as a function pointer comparison in both vm.c and the OB module.

Builds and tests pass on all platforms (GNU full/tiny, MSVC).
