+++ src/ph7/vm.c
@@ -64,16 +64,7 @@ struct VmRefObj
 	VmRefObj *pNext,*pPrev;               /* List of all referenced objects */
 };
 #define VM_REF_IDX_KEEP  0x001 /* Do not restore the memory object to the free list */
-/*
- * Output control buffer entry.
- * Refer to the implementation of [ob_start()] for more information.
- */
-typedef struct VmObEntry VmObEntry;
-struct VmObEntry
-{
-	ph7_value sCallback; /* User defined callback */
-	SyBlob sOB;          /* Output buffer consumer */
-};
+/* VmObEntry struct moved to ph7int.h */
 /*
  * Each installed shutdown callback (registered using [register_shutdown_function()] )
  * is stored in an instance of the following structure.
@@ -2112,7 +2103,6 @@ static sxi32 VmByteCodeDump(
 	return rc;
 }
 /* Forward declaration */
-static int VmObConsumer(const void *pData,unsigned int nDataLen,void *pUserData);
 static sxi32 VmUncaughtException(ph7_vm *pVm,ph7_class_instance *pThis);
 static sxi32 VmThrowException(ph7_vm *pVm,ph7_class_instance *pThis);
 static int VmMiniBacktrace(ph7_vm *pVm,SyBlob *pOut);
@@ -7162,438 +7152,7 @@ static int vm_builtin_get_defined_constants(ph7_context *pCtx,int nArg,ph7_value
 	ph7_result_value(pCtx,pArray);
 	return SXRET_OK;
 }
-/*
- * Section:
- *  Output Control (OB) functions.
- * Status:
- *    Stable.
- */
-/* Forward declaration */
-static void VmObRestore(ph7_vm *pVm,VmObEntry *pEntry);
-/*
- * void ob_clean(void)
- *  This function discards the contents of the output buffer.
- *  This function does not destroy the output buffer like ob_end_clean() does.
- * Parameter
- *  None
- * Return
- *  No value is returned.
- */
-static int vm_builtin_ob_clean(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	VmObEntry *pOb;
-	SXUNUSED(nArg); /* cc warning */
-	SXUNUSED(apArg);
-	/* Peek the top most OB */
-	pOb = (VmObEntry *)SySetPeek(&pVm->aOB);
-	if( pOb ){
-		SyBlobRelease(&pOb->sOB);
-	}
-	return PH7_OK;
-}
-/*
- * bool ob_end_clean(void)
- *  Clean (erase) the output buffer and turn off output buffering
- *  This function discards the contents of the topmost output buffer and turns
- *  off this output buffering. If you want to further process the buffer's contents
- *  you have to call ob_get_contents() before ob_end_clean() as the buffer contents
- *  are discarded when ob_end_clean() is called.
- * Parameter
- *  None
- * Return
- *  Returns TRUE on success or FALSE on failure. Reasons for failure are first that you called
- *  the function without an active buffer or that for some reason a buffer could not be deleted
- * (possible for special buffer)
- */
-static int vm_builtin_ob_end_clean(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	VmObEntry *pOb;
-	/* Pop the top most OB */
-	pOb = (VmObEntry *)SySetPop(&pVm->aOB);
-	if( pOb == 0){
-		/* No such OB,return FALSE */
-		ph7_result_bool(pCtx,0);
-		SXUNUSED(nArg); /* cc warning */
-		SXUNUSED(apArg);
-	}else{
-		/* Release */
-		VmObRestore(pVm,pOb);
-		/* Return true */
-		ph7_result_bool(pCtx,1);
-	}
-	return PH7_OK;
-}
-/*
- * string ob_get_contents(void)
- *  Gets the contents of the output buffer without clearing it.
- * Parameter
- *  None
- * Return
- *  This will return the contents of the output buffer or FALSE, if output buffering isn't active.
- */
-static int vm_builtin_ob_get_contents(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	VmObEntry *pOb;
-	/* Peek the top most OB */
-	pOb = (VmObEntry *)SySetPeek(&pVm->aOB);
-	if( pOb == 0 ){
-		/* No active OB,return FALSE */
-		ph7_result_bool(pCtx,0);
-		SXUNUSED(nArg); /* cc warning */
-		SXUNUSED(apArg);
-	}else{
-		/* Return contents */
-		ph7_result_string(pCtx,(const char *)SyBlobData(&pOb->sOB),(int)SyBlobLength(&pOb->sOB));
-	}
-	return PH7_OK;
-}
-/*
- * string ob_get_clean(void)
- * string ob_get_flush(void)
- *  Get current buffer contents and delete current output buffer.
- * Parameter
- *  None
- * Return
- *  This will return the contents of the output buffer or FALSE, if output buffering isn't active.
- */
-static int vm_builtin_ob_get_clean(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	VmObEntry *pOb;
-	/* Pop the top most OB */
-	pOb = (VmObEntry *)SySetPop(&pVm->aOB);
-	if( pOb == 0 ){
-		/* No active OB,return FALSE */
-		ph7_result_bool(pCtx,0);
-		SXUNUSED(nArg); /* cc warning */
-		SXUNUSED(apArg);
-	}else{
-		/* Return contents */
-		ph7_result_string(pCtx,(const char *)SyBlobData(&pOb->sOB),(int)SyBlobLength(&pOb->sOB)); /* Will make it's own copy */
-		/* Release */
-		VmObRestore(pVm,pOb);
-	}
-	return PH7_OK;
-}
-/*
- * int ob_get_length(void)
- *  Return the length of the output buffer.
- * Parameter
- *  None
- * Return
- *  Returns the length of the output buffer contents or FALSE if no buffering is active.
- */
-static int vm_builtin_ob_get_length(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	VmObEntry *pOb;
-	/* Peek the top most OB */
-	pOb = (VmObEntry *)SySetPeek(&pVm->aOB);
-	if( pOb == 0 ){
-		/* No active OB,return FALSE */
-		ph7_result_bool(pCtx,0);
-		SXUNUSED(nArg); /* cc warning */
-		SXUNUSED(apArg);
-	}else{
-		/* Return OB length */
-		ph7_result_int64(pCtx,(ph7_int64)SyBlobLength(&pOb->sOB));
-	}
-	return PH7_OK;
-}
-/*
- * int ob_get_level(void)
- *  Returns the nesting level of the output buffering mechanism.
- * Parameter
- *  None
- * Return
- *  Returns the level of nested output buffering handlers or zero if output buffering is not active.
- */
-static int vm_builtin_ob_get_level(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	int iNest;
-	SXUNUSED(nArg); /* cc warning */
-	SXUNUSED(apArg);
-	/* Nesting level */
-	iNest = (int)SySetUsed(&pVm->aOB);
-	/* Return the nesting value */
-	ph7_result_int(pCtx,iNest);
-	return PH7_OK;
-}
-/*
- * Output Buffer(OB) default VM consumer routine.All VM output is now redirected
- * to a stackable internal buffer,until the user call [ob_get_clean(),ob_end_clean(),...].
- * Refer to the implementation of [ob_start()] for more information.
- */
-static int VmObConsumer(const void *pData,unsigned int nDataLen,void *pUserData)
-{
-	ph7_vm *pVm = (ph7_vm *)pUserData;
-	VmObEntry *pEntry;
-	ph7_value sResult;
-	/* Peek the top most entry */
-	pEntry = (VmObEntry *)SySetPeek(&pVm->aOB);
-	if( pEntry == 0 ){
-		/* CAN'T HAPPEN */
-		return PH7_OK;
-	}
-	PH7_MemObjInit(pVm,&sResult);
-	if( ph7_value_is_callable(&pEntry->sCallback) && pVm->nObDepth < 15 ){
-		ph7_value sArg,*apArg[2];
-		/* Fill the first argument */
-		PH7_MemObjInitFromString(pVm,&sArg,0);
-		PH7_MemObjStringAppend(&sArg,(const char *)pData,nDataLen);
-		apArg[0] = &sArg;
-		/* Call the 'filter' callback */
-		pVm->nObDepth++;
-		PH7_VmCallUserFunction(pVm,&pEntry->sCallback,1,apArg,&sResult);
-		pVm->nObDepth--;
-		if( sResult.iFlags & MEMOBJ_STRING ){
-			/* Extract the function result */
-			pData = SyBlobData(&sResult.sBlob);
-			nDataLen = SyBlobLength(&sResult.sBlob);
-		}
-		PH7_MemObjRelease(&sArg);
-	}
-	if( nDataLen > 0 ){
-		/* Redirect the VM output to the internal buffer */
-		SyBlobAppend(&pEntry->sOB,pData,nDataLen);
-	}
-	/* Release */
-	PH7_MemObjRelease(&sResult);
-	return PH7_OK;
-}
-/*
- * Restore the default consumer.
- * Refer to the implementation of [ob_end_clean()] for more
- * information.
- */
-static void VmObRestore(ph7_vm *pVm,VmObEntry *pEntry)
-{
-	ph7_output_consumer *pCons = &pVm->sVmConsumer;
-	if( SySetUsed(&pVm->aOB) < 1 ){
-		/* No more stackable OB */
-		pCons->xConsumer = pCons->xDef;
-		pCons->pUserData = pCons->pDefData;
-	}
-	/* Release OB data */
-	PH7_MemObjRelease(&pEntry->sCallback);
-	SyBlobRelease(&pEntry->sOB);
-}
-/*
- * bool ob_start([ callback $output_callback] )
- * This function will turn output buffering on. While output buffering is active no output
- *  is sent from the script (other than headers), instead the output is stored in an internal
- *  buffer.
- * Parameter
- *  $output_callback
- *   An optional output_callback function may be specified. This function takes a string
- *   as a parameter and should return a string. The function will be called when the output
- *   buffer is flushed (sent) or cleaned (with ob_flush(), ob_clean() or similar function)
- *   or when the output buffer is flushed to the browser at the end of the request.
- *   When output_callback is called, it will receive the contents of the output buffer
- *   as its parameter and is expected to return a new output buffer as a result, which will
- *   be sent to the browser. If the output_callback is not a callable function, this function
- *   will return FALSE.
- *   If the callback function has two parameters, the second parameter is filled with
- *   a bit-field consisting of PHP_OUTPUT_HANDLER_START, PHP_OUTPUT_HANDLER_CONT
- *   and PHP_OUTPUT_HANDLER_END.
- *   If output_callback returns FALSE original input is sent to the browser.
- *   The output_callback parameter may be bypassed by passing a NULL value.
- * Return
- *   Returns TRUE on success or FALSE on failure.
- */
-static int vm_builtin_ob_start(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	VmObEntry sOb;
-	sxi32 rc;
-	/* Initialize the OB entry */
-	PH7_MemObjInit(pCtx->pVm,&sOb.sCallback);
-	SyBlobInit(&sOb.sOB,&pVm->sAllocator);
-	if( nArg > 0 && (apArg[0]->iFlags & (MEMOBJ_STRING|MEMOBJ_HASHMAP)) ){
-		/* Save the callback name for later invocation */
-		PH7_MemObjStore(apArg[0],&sOb.sCallback);
-	}
-	/* Push in the stack */
-	rc = SySetPut(&pVm->aOB,(const void *)&sOb);
-	if( rc != SXRET_OK ){
-		PH7_MemObjRelease(&sOb.sCallback);
-	}else{
-		ph7_output_consumer *pCons = &pVm->sVmConsumer;
-		/* Substitute the default VM consumer */
-		if( pCons->xConsumer != VmObConsumer ){
-			pCons->xDef = pCons->xConsumer;
-			pCons->pDefData = pCons->pUserData;
-			/* Install the new consumer */
-			pCons->xConsumer = VmObConsumer;
-			pCons->pUserData = pVm;
-		}
-	}
-	ph7_result_bool(pCtx,rc == SXRET_OK);
-	return PH7_OK;
-}
-/*
- * Flush Output buffer to the default VM output consumer.
- * Refer to the implementation of [ob_flush()] for more
- * information.
- */
-static sxi32 VmObFlush(ph7_vm *pVm,VmObEntry *pEntry,int bRelease)
-{
-	SyBlob *pBlob = &pEntry->sOB;
-	sxi32 rc;
-	/* Flush contents */
-	rc = PH7_OK;
-	if( SyBlobLength(pBlob) > 0 ){
-		/* Call the VM output consumer */
-		rc = pVm->sVmConsumer.xDef(SyBlobData(pBlob),SyBlobLength(pBlob),pVm->sVmConsumer.pDefData);
-		/* Increment VM output counter */
-		pVm->nOutputLen += SyBlobLength(pBlob);
-		if( rc != PH7_ABORT ){
-			rc = PH7_OK;
-		}
-	}
-	if( bRelease ){
-		VmObRestore(&(*pVm),pEntry);
-	}else{
-		/* Reset the blob */
-		SyBlobReset(pBlob);
-	}
-	return rc;
-}
-/*
- * void ob_flush(void)
- * void flush(void)
- *  Flush (send) the output buffer.
- * Parameter
- *  None
- * Return
- *  No return value.
- */
-static int vm_builtin_ob_flush(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	VmObEntry *pOb;
-	sxi32 rc;
-	/* Peek the top most OB entry */
-	pOb = (VmObEntry *)SySetPeek(&pVm->aOB);
-	if( pOb == 0 ){
-		/* Empty stack,return immediately */
-		SXUNUSED(nArg); /* cc warning */
-		SXUNUSED(apArg);
-		return PH7_OK;
-	}
-	/* Flush contents */
-	rc = VmObFlush(pVm,pOb,FALSE);
-	return rc;
-}
-/*
- * bool ob_end_flush(void)
- *  Flush (send) the output buffer and turn off output buffering.
- * Parameter
- *  None
- * Return
- *  Returns TRUE on success or FALSE on failure. Reasons for failure are first
- *  that you called the function without an active buffer or that for some reason
- *  a buffer could not be deleted (possible for special buffer).
- */
-static int vm_builtin_ob_end_flush(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	VmObEntry *pOb;
-	sxi32 rc;
-	/* Pop the top most OB entry */
-	pOb = (VmObEntry *)SySetPop(&pVm->aOB);
-	if( pOb == 0 ){
-		/* Empty stack,return FALSE */
-		ph7_result_bool(pCtx,0);
-		SXUNUSED(nArg); /* cc warning */
-		SXUNUSED(apArg);
-		return PH7_OK;
-	}
-	/* Flush contents */
-	rc = VmObFlush(pVm,pOb,TRUE);
-	/* Return true */
-	ph7_result_bool(pCtx,1);
-	return rc;
-}
-/*
- * void ob_implicit_flush([int $flag = true ])
- *  ob_implicit_flush() will turn implicit flushing on or off.
- *  Implicit flushing will result in a flush operation after every
- *  output call, so that explicit calls to flush() will no longer be needed.
- * Parameter
- *  $flag
- *   TRUE to turn implicit flushing on, FALSE otherwise.
- * Return
- *   Nothing
- */
-static int vm_builtin_ob_implicit_flush(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	/* NOTE: As of this version,this function is a no-op.
-	 * PH7 is smart enough to flush it's internal buffer when appropriate.
-	 */
-	SXUNUSED(pCtx);
-	SXUNUSED(nArg); /* cc warning */
-	SXUNUSED(apArg);
-	return PH7_OK;
-}
-/*
- * array ob_list_handlers(void)
- *  Lists all output handlers in use.
- * Parameter
- *  None
- * Return
- *  This will return an array with the output handlers in use (if any).
- */
-static int vm_builtin_ob_list_handlers(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	ph7_value *pArray;
-	VmObEntry *aEntry;
-	ph7_value sVal;
-	sxu32 n;
-	if( SySetUsed(&pVm->aOB) < 1 ){
-		/* Empty stack,return null */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	/* Create a new array */
-	pArray = ph7_context_new_array(pCtx);
-	if( pArray == 0 ){
-		/* Out of memory,return NULL */
-		SXUNUSED(nArg); /* cc warning */
-		SXUNUSED(apArg);
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	PH7_MemObjInit(pVm,&sVal);
-	/* Point to the installed OB entries */
-	aEntry = (VmObEntry *)SySetBasePtr(&pVm->aOB);
-	/* Perform the requested operation */
-	for( n = 0 ; n < SySetUsed(&pVm->aOB) ; n++ ){
-		VmObEntry *pEntry = &aEntry[n];
-		/* Extract handler name */
-		SyBlobReset(&sVal.sBlob);
-		if( pEntry->sCallback.iFlags & MEMOBJ_STRING ){
-			/* Callback,dup it's name */
-			SyBlobDup(&pEntry->sCallback.sBlob,&sVal.sBlob);
-		}else if( pEntry->sCallback.iFlags & MEMOBJ_HASHMAP ){
-			SyBlobAppend(&sVal.sBlob,"Class Method",sizeof("Class Method")-1);
-		}else{
-			SyBlobAppend(&sVal.sBlob,"default output handler",sizeof("default output handler")-1);
-		}
-		sVal.iFlags = MEMOBJ_STRING;
-		/* Perform the insertion */
-		ph7_array_add_elem(pArray,0/* Automatic index assign */,&sVal /* Will make it's own copy */);
-	}
-	PH7_MemObjRelease(&sVal);
-	/* Return the freshly created array */
-	ph7_result_value(pCtx,pArray);
-	return PH7_OK;
-}
+/* Output buffering builtins moved to vm_builtin_ob.c */
 /*
  * Section:
  *  Random numbers/string generators.
@@ -10323,331 +9882,7 @@ static int vm_builtin_require_once(ph7_context *pCtx,int nArg,ph7_value **apArg)
 	}
 	return SXRET_OK;
 }
-/*
- * Section:
- *  Command line arguments processing.
- * Status:
- *    Stable.
- */
-/*
- * Check if a short option argument [i.e: -c] is available in the command
- * line string. Return a pointer to the start of the stream on success.
- * NULL otherwise.
- */
-static const char * VmFindShortOpt(int c,const char *zIn,const char *zEnd)
-{
-	while( zIn < zEnd ){
-		if( zIn[0] == '-' && &zIn[1] < zEnd && (int)zIn[1] == c ){
-			/* Got one */
-			return &zIn[1];
-		}
-		/* Advance the cursor */
-		zIn++;
-	}
-	/* No such option */
-	return 0;
-}
-/*
- * Check if a long option argument [i.e: --opt] is available in the command
- * line string. Return a pointer to the start of the stream on success.
- * NULL otherwise.
- */
-static const char * VmFindLongOpt(const char *zLong,int nByte,const char *zIn,const char *zEnd)
-{
-	const char *zOpt;
-	while( zIn < zEnd ){
-		if( zIn[0] == '-' && &zIn[1] < zEnd && (int)zIn[1] == '-' ){
-			zIn += 2;
-			zOpt = zIn;
-			while( zIn < zEnd && !SyisSpace(zIn[0]) ){
-				if( zIn[0] == '=' /* --opt=val */){
-					break;
-				}
-				zIn++;
-			}
-			/* Test */
-			if( (int)(zIn-zOpt) == nByte && SyMemcmp(zOpt,zLong,nByte) == 0 ){
-				/* Got one,return it's value */
-				return zIn;
-			}
-
-		}else{
-			zIn++;
-		}
-	}
-	/* No such option */
-	return 0;
-}
-/*
- * Long option [i.e: --opt] arguments private data structure.
- */
-struct getopt_long_opt
-{
-	const char *zArgIn,*zArgEnd; /* Command line arguments */
-	ph7_value *pWorker;  /* Worker variable*/
-	ph7_value *pArray;   /* getopt() return value */
-	ph7_context *pCtx;   /* Call Context */
-};
-/* Forward declaration */
-static int VmProcessLongOpt(ph7_value *pKey,ph7_value *pValue,void *pUserData);
-/*
- * Extract short or long argument option values.
- */
-static void VmExtractOptArgValue(
-	ph7_value *pArray,  /* getopt() return value */
-	ph7_value *pWorker, /* Worker variable */
-	const char *zArg,   /* Argument stream */
-	const char *zArgEnd,/* End of the argument stream  */
-	int need_val,       /* TRUE to fetch option argument */
-	ph7_context *pCtx,  /* Call Context */
-	const char *zName   /* Option name */)
-{
-	ph7_value_bool(pWorker,0);
-	if( !need_val ){
-		/*
-		 * Option does not need arguments.
-		 * Insert the option name and a boolean FALSE.
-		 */
-		ph7_array_add_strkey_elem(pArray,(const char *)zName,pWorker); /* Will make it's own copy */
-	}else{
-		const char *zCur;
-		/* Extract option argument */
-		zArg++;
-		if( zArg < zArgEnd && zArg[0] == '=' ){
-			zArg++;
-		}
-		while( zArg < zArgEnd && (unsigned char)zArg[0] < 0xc0 && SyisSpace(zArg[0]) ){
-			zArg++;
-		}
-		if( zArg >= zArgEnd || zArg[0] == '-' ){
-			/*
-			 * Argument not found.
-			 * Insert the option name and a boolean FALSE.
-			 */
-			ph7_array_add_strkey_elem(pArray,(const char *)zName,pWorker); /* Will make it's own copy */
-			return;
-		}
-		/* Delimit the value */
-		zCur = zArg;
-		if( zArg[0] == '\'' || zArg[0] == '"' ){
-			int d = zArg[0];
-			/* Delimt the argument */
-			zArg++;
-			zCur = zArg;
-			while( zArg < zArgEnd ){
-				if( zArg[0] == d && zArg[-1] != '\\' ){
-					/* Delimiter found,exit the loop  */
-					break;
-				}
-				zArg++;
-			}
-			/* Save the value */
-			ph7_value_string(pWorker,zCur,(int)(zArg-zCur));
-			if( zArg < zArgEnd ){ zArg++; }
-		}else{
-			while( zArg < zArgEnd && !SyisSpace(zArg[0]) ){
-				zArg++;
-			}
-			/* Save the value */
-			ph7_value_string(pWorker,zCur,(int)(zArg-zCur));
-		}
-		/*
-		 * Check if we are dealing with multiple values.
-		 * If so,create an array to hold them,rather than a scalar variable.
-		 */
-		while( zArg < zArgEnd && (unsigned char)zArg[0] < 0xc0 && SyisSpace(zArg[0]) ){
-			zArg++;
-		}
-		if( zArg < zArgEnd && zArg[0] != '-' ){
-			ph7_value *pOptArg; /* Array of option arguments */
-			pOptArg = ph7_context_new_array(pCtx);
-			if( pOptArg == 0 ){
-				ph7_context_throw_error(pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-			}else{
-				/* Insert the first value */
-				ph7_array_add_elem(pOptArg,0,pWorker); /* Will make it's own copy */
-				for(;;){
-					if( zArg >= zArgEnd || zArg[0] == '-' ){
-						/* No more value */
-						break;
-					}
-					/* Delimit the value */
-					zCur = zArg;
-					if( zArg < zArgEnd && zArg[0] == '\\' ){
-						zArg++;
-						zCur = zArg;
-					}
-					while( zArg < zArgEnd && !SyisSpace(zArg[0]) ){
-						zArg++;
-					}
-					/* Reset the string cursor */
-					ph7_value_reset_string_cursor(pWorker);
-					/* Save the value */
-					ph7_value_string(pWorker,zCur,(int)(zArg-zCur));
-					/* Insert */
-					ph7_array_add_elem(pOptArg,0,pWorker); /* Will make it's own copy */
-					/* Jump trailing white spaces */
-					while( zArg < zArgEnd && (unsigned char)zArg[0] < 0xc0 && SyisSpace(zArg[0]) ){
-						zArg++;
-					}
-				}
-				/* Insert the option arg array */
-				ph7_array_add_strkey_elem(pArray,(const char *)zName,pOptArg); /* Will make it's own copy */
-				/* Safely release */
-				ph7_context_release_value(pCtx,pOptArg);
-			}
-		}else{
-			/* Single value */
-			ph7_array_add_strkey_elem(pArray,(const char *)zName,pWorker); /* Will make it's own copy */
-		}
-	}
-}
-/*
- * array getopt(string $options[,array $longopts ])
- *   Gets options from the command line argument list.
- * Parameters
- *  $options
- *   Each character in this string will be used as option characters
- *   and matched against options passed to the script starting with
- *   a single hyphen (-). For example, an option string "x" recognizes
- *   an option -x. Only a-z, A-Z and 0-9 are allowed.
- *  $longopts
- *   An array of options. Each element in this array will be used as option
- *   strings and matched against options passed to the script starting with
- *   two hyphens (--). For example, an longopts element "opt" recognizes an
- *   option --opt.
- * Return
- *  This function will return an array of option / argument pairs or FALSE
- *  on failure.
- */
-static int vm_builtin_getopt(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	const char *zIn,*zEnd,*zArg,*zArgIn,*zArgEnd;
-	struct getopt_long_opt sLong;
-	ph7_value *pArray,*pWorker;
-	SyBlob *pArg;
-	int nByte;
-	if( nArg < 1 || !ph7_value_is_string(apArg[0]) ){
-		/* Missing/Invalid arguments,return FALSE */
-		ph7_context_throw_error(pCtx,PH7_CTX_ERR,"Missing/Invalid option arguments");
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Extract option arguments */
-	zIn  = ph7_value_to_string(apArg[0],&nByte);
-	zEnd = &zIn[nByte];
-	/* Point to the string representation of the $argv[] array */
-	pArg = &pCtx->pVm->sArgv;
-	/* Create a new empty array and a worker variable */
-	pArray = ph7_context_new_array(pCtx);
-	pWorker = ph7_context_new_scalar(pCtx);
-	if( pArray == 0 || pWorker == 0 ){
-		ph7_context_throw_error(pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( SyBlobLength(pArg) < 1 ){
-		/* Empty command line,return the empty array*/
-		ph7_result_value(pCtx,pArray);
-		/* Everything will be released automatically when we return
-		 * from this function.
-		 */
-		return PH7_OK;
-	}
-	zArgIn = (const char *)SyBlobData(pArg);
-	zArgEnd = &zArgIn[SyBlobLength(pArg)];
-	/* Fill the long option structure */
-	sLong.pArray = pArray;
-	sLong.pWorker = pWorker;
-	sLong.zArgIn =  zArgIn;
-	sLong.zArgEnd = zArgEnd;
-	sLong.pCtx = pCtx;
-	/* Start processing */
-	while( zIn < zEnd ){
-		int c = zIn[0];
-		int need_val = 0;
-		/* Advance the stream cursor */
-		zIn++;
-		/* Ignore non-alphanum characters */
-		if( !SyisAlphaNum(c) ){
-			continue;
-		}
-		if( zIn < zEnd && zIn[0] == ':' ){
-			zIn++;
-			need_val = 1;
-			if( zIn < zEnd && zIn[0] == ':' ){
-				zIn++;
-			}
-		}
-		/* Find option */
-		zArg = VmFindShortOpt(c,zArgIn,zArgEnd);
-		if( zArg == 0 ){
-			/* No such option */
-			continue;
-		}
-		/* Extract option argument value */
-		VmExtractOptArgValue(pArray,pWorker,zArg,zArgEnd,need_val,pCtx,(const char *)&c);
-	}
-	if( nArg > 1 && ph7_value_is_array(apArg[1]) && ph7_array_count(apArg[1]) > 0 ){
-		/* Process long options */
-		ph7_array_walk(apArg[1],VmProcessLongOpt,&sLong);
-	}
-	/* Return the option array */
-	ph7_result_value(pCtx,pArray);
-	/*
-	 * Don't worry about freeing memory, everything will be released
-	 * automatically as soon we return from this foreign function.
-	 */
-	return PH7_OK;
-}
-/*
- * Array walker callback used for processing long options values.
- */
-static int VmProcessLongOpt(ph7_value *pKey,ph7_value *pValue,void *pUserData)
-{
-	struct getopt_long_opt *pOpt = (struct getopt_long_opt *)pUserData;
-	const char *zArg,*zOpt,*zEnd;
-	int need_value = 0;
-	int nByte;
-	/* Value must be of type string */
-	if( !ph7_value_is_string(pValue) ){
-		/* Simply ignore */
-		return PH7_OK;
-	}
-	zOpt = ph7_value_to_string(pValue,&nByte);
-	if( nByte < 1 ){
-		/* Empty string,ignore */
-		return PH7_OK;
-	}
-	zEnd = &zOpt[nByte - 1];
-	if( zEnd[0] == ':' ){
-		char *zTerm;
-		/* Try to extract a value */
-		need_value = 1;
-		while( zEnd >= zOpt && zEnd[0] == ':' ){
-			zEnd--;
-		}
-		if( zOpt >= zEnd ){
-			/* Empty string,ignore */
-			SXUNUSED(pKey);
-			return PH7_OK;
-		}
-		zEnd++;
-		zTerm = (char *)zEnd;
-		zTerm[0] = 0;
-	}else{
-		zEnd = &zOpt[nByte];
-	}
-	/* Find the option */
-	zArg = VmFindLongOpt(zOpt,(int)(zEnd-zOpt),pOpt->zArgIn,pOpt->zArgEnd);
-	if( zArg == 0 ){
-		/* No such option,return immediately */
-		return PH7_OK;
-	}
-	/* Try to extract a value */
-	VmExtractOptArgValue(pOpt->pArray,pOpt->pWorker,zArg,pOpt->zArgEnd,need_value,pOpt->pCtx,zOpt);
-	return PH7_OK;
-}
+/* Getopt builtins moved to vm_builtin_getopt.c */
 /* JSON encoding/decoding routines moved to vm_json.c */
 /* XML processing and UTF-8 routines moved to vm_xml.c */
 /* Table of built-in VM functions. */
